[tmp] safe cast to remove warning

+++ src/nvim/normal.c
@@ -845,8 +845,10 @@ getcount:
        * because if it's put back with vungetc() it's too late to apply
        * mapping. */
       no_mapping--;
-      while (enc_utf8 && lang && (c = vpeekc()) > 0
-             && (c >= 0x100 || mb_byte2len(c) > 1)) {
+      while (enc_utf8
+             && lang
+             && (c = vpeekc()) > 0
+             && (c >= 0x100 || mb_byte2len((uint8_t)c) > 1)) {
         c = plain_vgetc();
         if (!utf_iscomposing(c)) {
           vungetc(c);                   /* it wasn't, put it back */
